backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
google-generativeai==0.3.1
python-multipart==0.0.6
alembic==1.12.1
pandas==2.1.3
numpy==1.26.2
Pillow==10.2.0

